+++ source/MRDotNet2Test/src/MeshCollideTests.cs
@@ -16,7 +16,7 @@ namespace MRTest
             var conv = getVectorConverters(meshA, meshB);
 
             var intersections = findCollidingEdgeTrisPrecise(meshA, meshB, conv.toInt);
-            Assert.That(intersections.size(), Is.EqualTo(152));
+            Assert.That(intersections.size(), Is.EqualTo((nuint)152));
             var edgeATriBCount = 0;
             var edgeBTriACount = 0;
             for (nuint i = 0; i < intersections.size(); i++)
@@ -31,17 +31,17 @@ namespace MRTest
 
             var contours = orderIntersectionContours(meshA.topology, meshB.topology, intersections);
 
-            Assert.That(contours.size(), Is.EqualTo(4));
-            Assert.That(contours[0].size(), Is.EqualTo(71));
-            Assert.That(contours[1].size(), Is.EqualTo(7));
-            Assert.That(contours[2].size(), Is.EqualTo(69));
-            Assert.That(contours[3].size(), Is.EqualTo(9));
+            Assert.That(contours.size(), Is.EqualTo((nuint)4));
+            Assert.That(contours[0].size(), Is.EqualTo((nuint)71));
+            Assert.That(contours[1].size(), Is.EqualTo((nuint)7));
+            Assert.That(contours[2].size(), Is.EqualTo((nuint)69));
+            Assert.That(contours[3].size(), Is.EqualTo((nuint)9));
             var aConts = new Std.Vector_MROneMeshContour();
             getOneMeshIntersectionContours(meshA, meshB, contours, aConts, null, conv);
-            Assert.That(aConts.size(), Is.EqualTo(4));
+            Assert.That(aConts.size(), Is.EqualTo((nuint)4));
             var bConts = new Std.Vector_MROneMeshContour();
             getOneMeshIntersectionContours(meshA, meshB, contours, null, bConts, conv);
-            Assert.That(bConts.size(), Is.EqualTo(4));
+            Assert.That(bConts.size(), Is.EqualTo((nuint)4));
 
             nuint posCount = 0;
             for (nuint i = 0; i < aConts.size(); i++)
@@ -49,7 +49,7 @@ namespace MRTest
                 posCount += aConts[i].intersections.size();
             }
 
-            Assert.That(posCount, Is.EqualTo(156) );
+            Assert.That(posCount, Is.EqualTo((nuint)156) );
 
         }
